UploadWebsite.bat: minor improvements
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@23397 154be670-7969-47af-9ab0-c6fb378d2295

+++ www/UploadWebsite.bat
@@ -6,7 +6,7 @@ rem variable to point to your scp program the PLINK variable to the plink
 rem program and the ZIP variable to your zip program
 
 rem example docserverlogin.bat file
-rem 
+rem
 rem @echo off
 rem set USERNAME=myname
 rem set PASSWORD=mypassword
@@ -14,17 +14,16 @@ rem set PSCP="C:\Programme\PuttY\pscp.exe"
 rem set PLINK="C:\Programme\Putty\plink.exe"
 rem set ZIP="C:\Programme\7-zip\7z.exe"
 
-
-rd /s /q ..\..\tsvnsite
-
-cd scripts
+if exist ..\..\tsvnsite rd /s /q ..\..\tsvnsite
+pushd scripts
 python generatesite.py
-cd ..
+popd
 
 if exist ..\doc\docserverlogin.bat call ..\doc\docserverlogin.bat
 if exist ..\serverlogin.bat call ..\serverlogin.bat
 
 pushd ..\..\tsvnsite
+if exist website.zip del website.zip
 
 %ZIP% a -r -tzip website.zip * > NUL
